// packages/companion-module-base/src/host-api/api.ts
export type EncodeIsVisible<T extends CompanionInputFieldBase> = Omit<T, 'isVisible' | 'isVisibleExpression'> & {
	isVisibleFn?: string
	isVisibleFnType?: 'function' | 'expression'
}